Fix image loading issues by switching from MHTML to HTML format
Images hidden behind buttons were not displayed correctly when importing in MHTML. Sometimes, only the first image was displayed correctly, while the others remained white or were broken.

To solve this problem, we switched to importing in HTML, which allows for better loading and display of images. This method ensures that all images are loaded and displayed correctly, improving the user experience.